fix(SetupApi): abstract interceptors to child classes
