+++ src/native/index.ts
@@ -1,6 +1,16 @@
 import { XMLHttpRequestInterceptor } from '@mswjs/interceptors/lib/interceptors/XMLHttpRequest'
-import { createSetupServer } from '../node/createSetupServer'
+import { RequestHandler } from '../handlers/RequestHandler'
+import { SetupServerApi } from '../node/SetupServerApi'
 
-// Provision request interception via patching the `XMLHttpRequest` class only
-// in React Native. There is no `http`/`https` modules in that environment.
-export const setupServer = createSetupServer(XMLHttpRequestInterceptor)
+/**
+ * Sets up a requests interception in React Native with the given request handlers.
+ * @param {RequestHandler[]} handlers List of request handlers.
+ * @see {@link https://mswjs.io/docs/api/setup-server `setupServer`}
+ */
+export function setupServer(
+  ...handlers: Array<RequestHandler>
+): SetupServerApi {
+  // Provision request interception via patching the `XMLHttpRequest` class only
+  // in React Native. There is no `http`/`https` modules in that environment.
+  return new SetupServerApi([XMLHttpRequestInterceptor], handlers)
+}
