+++ src/node/index.ts
@@ -1,2 +1,3 @@
-export { setupServer, ServerLifecycleEventsMap } from './setupServer'
+export { ServerLifecycleEventsMap } from './SetupServerApi'
+export { setupServer } from './setupServer'
 export type { SetupServerApi } from './glossary'
