+++ test/msw-api/setup-worker/input-validation.test.ts
@@ -16,7 +16,7 @@ test('throws an error given an Array of request handlers to "setupWorker"', asyn
   expect(exceptions).toEqual(
     expect.arrayContaining([
       expect.stringContaining(
-        '[MSW] Failed to call "SetupWorkerApi" given an Array of request handlers (SetupWorkerApi([a, b])), expected to receive each handler individually: SetupWorkerApi(a, b).',
+        '[MSW] Failed to call "SetupWorkerApi" given an Array of request handlers. Make sure you spread the request handlers when calling this function.',
       ),
     ]),
   )
